Rename all source files to camelCase. Consistent with the standard naming I've noticed since starting this project
