+++ src/metadata_browser/lightningComponentBundle.ts
