+++ src/metadata_browser/metadataComponent.ts
