+++ src/metadata_browser/sfdxProjectBrowser.ts
@@ -1,11 +1,11 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { SfdxProject } from '@salesforce/core';
-import { CustomField } from './CustomField';
-import { EmailToCaseSettings } from './EmailToCaseSettings';
-import { LightningComponentBundle } from './LightningComponentBundle';
-import { PicklistField } from './PicklistField';
-import { RecordType } from './RecordType';
+import { CustomField } from './customField';
+import { EmailToCaseSettings } from './emailToCaseSettings';
+import { LightningComponentBundle } from './lightningComponentBundle';
+import { PicklistField } from './picklistField';
+import { RecordType } from './recordType';
 
 // Tools for browsing/navigating the metadata in an SFDX project
 export class SfdxProjectBrowser {
